requirements.txt
pandas==2.1.4
xlrd==2.0.1
pyarrow==11.0.0
beautifulsoup4==4.12.3
requests==2.28.0
pyyaml==6.0